migrate permissions cache
