+++ packages/twenty-server/src/engine/api/rest/core/handlers/rest-api-base.handler.ts
@@ -33,8 +33,8 @@ import {
   PermissionsExceptionMessage,
 } from 'src/engine/metadata-modules/permissions/permissions.exception';
 import { UserRoleService } from 'src/engine/metadata-modules/user-role/user-role.service';
-import { WorkspacePermissionsCacheService } from 'src/engine/metadata-modules/workspace-permissions-cache/workspace-permissions-cache.service';
 import { TwentyORMManager } from 'src/engine/twenty-orm/twenty-orm.manager';
+import { WorkspaceCacheService } from 'src/engine/workspace-cache/services/workspace-cache.service';
 import { WorkspaceCacheStorageService } from 'src/engine/workspace-cache-storage/workspace-cache-storage.service';
 import { standardObjectMetadataDefinitions } from 'src/engine/workspace-manager/workspace-sync-metadata/standard-objects';
 import { shouldExcludeFromWorkspaceApi } from 'src/engine/workspace-manager/workspace-sync-metadata/utils/should-exclude-from-workspace-api.util';
@@ -58,7 +58,7 @@ export abstract class RestApiBaseHandler {
   @Inject()
   protected readonly twentyORMManager: TwentyORMManager;
   @Inject()
-  protected readonly workspacePermissionsCacheService: WorkspacePermissionsCacheService;
+  protected readonly workspaceCacheService: WorkspaceCacheService;
   @Inject()
   protected readonly createdByFromAuthContextService: CreatedByFromAuthContextService;
   @Inject()
@@ -115,15 +115,13 @@ export abstract class RestApiBaseHandler {
       roleId = userWorkspaceRoleId;
     }
 
-    const objectMetadataPermissions =
-      await this.workspacePermissionsCacheService.getObjectRecordPermissionsForRoles(
-        {
-          workspaceId: authContext.workspace.id,
-          roleIds: [roleId],
-        },
+    const { rolesPermissions } =
+      await this.workspaceCacheService.getOrRecompute(
+        authContext.workspace.id,
+        ['rolesPermissions'],
       );
 
-    return { objectsPermissions: objectMetadataPermissions[roleId] };
+    return { objectsPermissions: rolesPermissions[roleId] };
   };
 
   async computeSelectedFields({
