+++ packages/twenty-server/src/engine/metadata-modules/permission-flag/permission-flag.module.ts
@@ -5,11 +5,13 @@ import { PermissionFlagEntity } from 'src/engine/metadata-modules/permission-fla
 import { PermissionFlagService } from 'src/engine/metadata-modules/permission-flag/permission-flag.service';
 import { RoleEntity } from 'src/engine/metadata-modules/role/role.entity';
 import { WorkspacePermissionsCacheModule } from 'src/engine/metadata-modules/workspace-permissions-cache/workspace-permissions-cache.module';
+import { WorkspaceCacheModule } from 'src/engine/workspace-cache/workspace-cache.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([PermissionFlagEntity, RoleEntity]),
     WorkspacePermissionsCacheModule,
+    WorkspaceCacheModule,
   ],
 
   providers: [PermissionFlagService],
