+++ packages/twenty-server/src/engine/metadata-modules/permission-flag/permission-flag.service.ts
@@ -13,7 +13,7 @@ import {
   PermissionsExceptionMessage,
 } from 'src/engine/metadata-modules/permissions/permissions.exception';
 import { RoleEntity } from 'src/engine/metadata-modules/role/role.entity';
-import { WorkspacePermissionsCacheService } from 'src/engine/metadata-modules/workspace-permissions-cache/workspace-permissions-cache.service';
+import { WorkspaceCacheService } from 'src/engine/workspace-cache/services/workspace-cache.service';
 
 export class PermissionFlagService {
   constructor(
@@ -21,7 +21,7 @@ export class PermissionFlagService {
     private readonly roleRepository: Repository<RoleEntity>,
     @InjectDataSource()
     private readonly coreDataSource: DataSource,
-    private readonly workspacePermissionsCacheService: WorkspacePermissionsCacheService,
+    private readonly workspaceCacheService: WorkspaceCacheService,
   ) {}
 
   public async upsertPermissionFlags({
@@ -128,12 +128,9 @@ export class PermissionFlagService {
     } finally {
       await queryRunner.release();
 
-      await this.workspacePermissionsCacheService.recomputeRolesPermissionsCache(
-        {
-          workspaceId,
-          roleIds: [input.roleId],
-        },
-      );
+      await this.workspaceCacheService.invalidate(workspaceId, [
+        'rolesPermissions',
+      ]);
     }
   }
