+++ packages/twenty-server/src/engine/metadata-modules/user-role/user-role.module.ts
@@ -6,12 +6,14 @@ import { RoleTargetsEntity } from 'src/engine/metadata-modules/role/role-targets
 import { RoleEntity } from 'src/engine/metadata-modules/role/role.entity';
 import { UserRoleService } from 'src/engine/metadata-modules/user-role/user-role.service';
 import { WorkspacePermissionsCacheModule } from 'src/engine/metadata-modules/workspace-permissions-cache/workspace-permissions-cache.module';
+import { WorkspaceCacheModule } from 'src/engine/workspace-cache/workspace-cache.module';
 
 @Module({
   imports: [
     TypeOrmModule.forFeature([RoleEntity, RoleTargetsEntity]),
     TypeOrmModule.forFeature([UserWorkspaceEntity]),
     WorkspacePermissionsCacheModule,
+    WorkspaceCacheModule,
   ],
   providers: [UserRoleService],
   exports: [UserRoleService],
