--- packages/twenty-server/src/engine/metadata-modules/workspace-permissions-cache/workspace-permissions-cache-storage.service.ts
@@ -1,167 +1,0 @@
-import { Injectable, Logger } from '@nestjs/common';
-
-import { type ObjectsPermissionsByRoleId } from 'twenty-shared/types';
-import { v4 } from 'uuid';
-
-import { InjectCacheStorage } from 'src/engine/core-modules/cache-storage/decorators/cache-storage.decorator';
-import { CacheStorageService } from 'src/engine/core-modules/cache-storage/services/cache-storage.service';
-import { CacheStorageNamespace } from 'src/engine/core-modules/cache-storage/types/cache-storage-namespace.enum';
-import { type UserWorkspaceRoleMap } from 'src/engine/metadata-modules/workspace-permissions-cache/types/user-workspace-role-map.type';
-import { WORKSPACE_CACHE_KEYS } from 'src/engine/workspace-cache-storage/workspace-cache-storage.service';
-
-const TTL_INFINITE = 0;
-
-@Injectable()
-export class WorkspacePermissionsCacheStorageService {
-  logger = new Logger(WorkspacePermissionsCacheStorageService.name);
-
-  constructor(
-    @InjectCacheStorage(CacheStorageNamespace.EngineWorkspace)
-    private readonly cacheStorageService: CacheStorageService,
-  ) {}
-
-  async setRolesPermissionsVersion(workspaceId: string): Promise<string> {
-    const rolesPermissionsVersion = v4();
-
-    await this.cacheStorageService.set<string>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsRolesPermissionsVersion}:${workspaceId}`,
-      rolesPermissionsVersion,
-      TTL_INFINITE,
-    );
-
-    return rolesPermissionsVersion;
-  }
-
-  async setRolesPermissions(
-    workspaceId: string,
-    permissions: ObjectsPermissionsByRoleId,
-  ): Promise<{
-    newRolesPermissionsVersion: string;
-  }> {
-    const [, newRolesPermissionsVersion] = await Promise.all([
-      this.cacheStorageService.set<ObjectsPermissionsByRoleId>(
-        `${WORKSPACE_CACHE_KEYS.MetadataPermissionsRolesPermissions}:${workspaceId}`,
-        permissions,
-        TTL_INFINITE,
-      ),
-      this.setRolesPermissionsVersion(workspaceId),
-    ]);
-
-    return { newRolesPermissionsVersion };
-  }
-
-  getRolesPermissions(
-    workspaceId: string,
-  ): Promise<ObjectsPermissionsByRoleId | undefined> {
-    return this.cacheStorageService.get<ObjectsPermissionsByRoleId>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsRolesPermissions}:${workspaceId}`,
-    );
-  }
-
-  getRolesPermissionsVersion(workspaceId: string): Promise<string | undefined> {
-    return this.cacheStorageService.get<string>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsRolesPermissionsVersion}:${workspaceId}`,
-    );
-  }
-
-  async setUserWorkspaceRoleMap(
-    workspaceId: string,
-    userWorkspaceRoleMap: UserWorkspaceRoleMap,
-  ): Promise<void> {
-    await Promise.all([
-      this.cacheStorageService.set<UserWorkspaceRoleMap>(
-        `${WORKSPACE_CACHE_KEYS.MetadataPermissionsUserWorkspaceRoleMap}:${workspaceId}`,
-        userWorkspaceRoleMap,
-        TTL_INFINITE,
-      ),
-      this.setUserWorkspaceRoleMapVersion(workspaceId),
-    ]);
-  }
-
-  async setUserWorkspaceRoleMapVersion(workspaceId: string) {
-    const userWorkspaceRoleMapVersion = v4();
-
-    await this.cacheStorageService.set<string>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsUserWorkspaceRoleMapVersion}:${workspaceId}`,
-      userWorkspaceRoleMapVersion,
-      TTL_INFINITE,
-    );
-
-    return userWorkspaceRoleMapVersion;
-  }
-
-  getUserWorkspaceRoleMap(
-    workspaceId: string,
-  ): Promise<Record<string, string> | undefined> {
-    return this.cacheStorageService.get<Record<string, string>>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsUserWorkspaceRoleMap}:${workspaceId}`,
-    );
-  }
-
-  getUserWorkspaceRoleMapVersion(
-    workspaceId: string,
-  ): Promise<string | undefined> {
-    return this.cacheStorageService.get<string>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsUserWorkspaceRoleMapVersion}:${workspaceId}`,
-    );
-  }
-
-  removeUserWorkspaceRoleMap(workspaceId: string) {
-    return this.cacheStorageService.del(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsUserWorkspaceRoleMap}:${workspaceId}`,
-    );
-  }
-
-  async setApiKeyRoleMap(
-    workspaceId: string,
-    apiKeyRoleMap: Record<string, string>,
-  ): Promise<void> {
-    await Promise.all([
-      this.cacheStorageService.set<Record<string, string>>(
-        `${WORKSPACE_CACHE_KEYS.MetadataPermissionsApiKeyRoleMap}:${workspaceId}`,
-        apiKeyRoleMap,
-        TTL_INFINITE,
-      ),
-      this.setApiKeyRoleMapVersion(workspaceId),
-    ]);
-  }
-
-  async getApiKeyRoleMap(
-    workspaceId: string,
-  ): Promise<Record<string, string> | undefined> {
-    return this.cacheStorageService.get<Record<string, string>>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsApiKeyRoleMap}:${workspaceId}`,
-    );
-  }
-
-  async getApiKeyRoleMapVersion(
-    workspaceId: string,
-  ): Promise<string | undefined> {
-    return this.cacheStorageService.get<string>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsApiKeyRoleMapVersion}:${workspaceId}`,
-    );
-  }
-
-  async removeApiKeyRoleMap(workspaceId: string): Promise<void> {
-    await Promise.all([
-      this.cacheStorageService.del(
-        `${WORKSPACE_CACHE_KEYS.MetadataPermissionsApiKeyRoleMap}:${workspaceId}`,
-      ),
-      this.cacheStorageService.del(
-        `${WORKSPACE_CACHE_KEYS.MetadataPermissionsApiKeyRoleMapVersion}:${workspaceId}`,
-      ),
-    ]);
-  }
-
-  private async setApiKeyRoleMapVersion(workspaceId: string) {
-    const apiKeyRoleMapVersion = v4();
-
-    await this.cacheStorageService.set<string>(
-      `${WORKSPACE_CACHE_KEYS.MetadataPermissionsApiKeyRoleMapVersion}:${workspaceId}`,
-      apiKeyRoleMapVersion,
-      TTL_INFINITE,
-    );
-
-    return apiKeyRoleMapVersion;
-  }
-}
