+++ packages/twenty-server/src/engine/workspace-manager/dev-seeder/core/services/dev-seeder-permissions.service.ts
@@ -12,7 +12,7 @@ import { ObjectPermissionService } from 'src/engine/metadata-modules/object-perm
 import { RoleEntity } from 'src/engine/metadata-modules/role/role.entity';
 import { RoleService } from 'src/engine/metadata-modules/role/role.service';
 import { UserRoleService } from 'src/engine/metadata-modules/user-role/user-role.service';
-import { WorkspacePermissionsCacheService } from 'src/engine/metadata-modules/workspace-permissions-cache/workspace-permissions-cache.service';
+import { WorkspaceCacheService } from 'src/engine/workspace-cache/services/workspace-cache.service';
 import {
   SEED_APPLE_WORKSPACE_ID,
   SEED_YCOMBINATOR_WORKSPACE_ID,
@@ -36,7 +36,7 @@ export class DevSeederPermissionsService {
     private readonly objectMetadataRepository: Repository<ObjectMetadataEntity>,
     @InjectRepository(RoleEntity)
     private readonly roleRepository: Repository<RoleEntity>,
-    private readonly workspacePermissionsCacheService: WorkspacePermissionsCacheService,
+    private readonly workspaceCacheService: WorkspaceCacheService,
     private readonly fieldPermissionService: FieldPermissionService,
     @InjectDataSource()
     private readonly coreDataSource: DataSource,
@@ -77,14 +77,10 @@ export class DevSeederPermissionsService {
         ])
         .execute();
 
-      await this.workspacePermissionsCacheService.recomputeApiKeyRoleMapCache({
-        workspaceId,
-      });
-      await this.workspacePermissionsCacheService.recomputeUserWorkspaceRoleMapCache(
-        {
-          workspaceId,
-        },
-      );
+      await this.workspaceCacheService.invalidate(workspaceId, [
+        'apiKeyRoleMap',
+        'userWorkspaceRoleMap',
+      ]);
     } catch (error) {
       this.logger.error(
         `Could not assign role to test API key: ${error.message}`,
